JS now asking for delete comments\lots

+++ app/View/Views/showuserscomments.php
@@ -10,8 +10,17 @@
             <td><?= $comment['description'] ?></td>
             <td><?= $comment['add_time'] ?></td>
             <td><a href="/changecomment/<?= $comment['id'] ?>">Редактировать</a><br>
-            <a href="/deletecomment/<?= $comment['id'] ?>">Удалить</a></td>
+            <a onclick="confirmDeleteComment()" href="/deletecomment/<?= $comment['id'] ?>">Удалить</a></td>
         </tr>
         <?php endforeach; ?>
     </table>
 </div>
+<script type="text/javascript">
+    function confirmDeleteComment() {
+        if (confirm("Вы уверены, что хотите удалить комментарий?")) {
+            return true;
+        } else {
+            event.preventDefault();
+        }
+    }
+</script>
